[refactor]: Refactor exercise table
