+++ src/components/exercise/post/Badge.tsx
